Fixed Fisheye no ugly walls

+++ srcs/bresenham.c
@@ -6,7 +6,7 @@
 /*   By: mratsima <mratsima@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/24 13:17:03 by fsamy-an          #+#    #+#             */
-/*   Updated: 2025/11/11 09:55:23 by mratsima         ###   ########.fr       */
+/*   Updated: 2025/11/13 14:00:03 by mratsima         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ t_point calculate_end(t_point begin, float angle, int max_distance)
 	return (end);
 }
 
-void draw_wall_lines(t_display *display, int distance, int pixel_index)
+void draw_wall_lines(t_display *display, float distance, int pixel_index)
 {
 	t_line *line;
 	int line_size;
@@ -69,7 +69,7 @@ void	cast_ray(t_point begin,t_display *display, int d)
 	t_point true_end;
 	float	angle;
 	int pixel_index;
-	int distance;
+	float distance;
 
 	/*
 		Dans cette fonction nous mettons angle = -FOV/2 pour que la direction principale se trouve au milieu
